refactoring Image related code
