+++ waste_watcher_adafruitIO/include/Image.h
@@ -1,0 +1,10 @@
+#ifndef IMAGE 
+#define IMAGE
+
+#include <Arduino.h>
+
+void configInitCamera(bool debug);
+String takePhoto(bool debug);
+
+
+#endif
