--- waste_watcher_adafruitIO/lib/Fullness/Fullness.h
@@ -1,19 +1,0 @@
-#ifndef FULLNESS 
-#define FULLNESS
-
-float distanceRead(
-  int ultrasonicTrigPin = 13,
-  int ultrasonicEchoPin = 12
-);
-
-int calculateFullness(
-  int ultrasonicTrigPin = 13,
-  int ultrasonicEchoPin = 12,
-  int binHeight = 50, 
-  int minValue = 0,
-  int maxValue = 50,
-  int retries = 5
-);
-
-
-#endif
