+++ waste_watcher_adafruitIO/src/template-config.h
@@ -19,9 +19,6 @@
 #define IO_KEY        "<YOUR_ADAFRUIT_IO_KEY_HERE>"
 #define FEEDNAME      "<YOUR_ADAFRUIT_IO_FEEDNAME>"
 
-AdafruitIO_WiFi io(IO_USERNAME, IO_KEY, WIFI_SSID, WIFI_PASS);
-
-
 // ultrasonic sensor pins
 const int ultrasonicTrigPin = 13;
 const int ultrasonicEchoPin = 12;
@@ -31,8 +28,31 @@ const int ultrasonicEchoPin = 12;
 #define TIME_TO_SLEEP 1800   // TIME ESP32 will sleep for 1800 sec
 
 // system variables
-bool debug = true; // set false, if not debugging
 const int binHeight = 50; // units: cm, used to calculate fullness
 const int retries = 5; // # of retries for collecting fullness
+
+// camera 
+// === camera pin definitions ===
+const int flashLED = 4;
+
+//CAMERA_MODEL_AI_THINKER
+#define PWDN_GPIO_NUM     32
+#define RESET_GPIO_NUM    -1
+#define XCLK_GPIO_NUM      0
+#define SIOD_GPIO_NUM     26
+#define SIOC_GPIO_NUM     27
+
+#define Y9_GPIO_NUM       35
+#define Y8_GPIO_NUM       34
+#define Y7_GPIO_NUM       39
+#define Y6_GPIO_NUM       36
+#define Y5_GPIO_NUM       21
+#define Y4_GPIO_NUM       19
+#define Y3_GPIO_NUM       18
+#define Y2_GPIO_NUM        5
+#define VSYNC_GPIO_NUM    25
+#define HREF_GPIO_NUM     23
+#define PCLK_GPIO_NUM     22
+
 
 #endif
